Created utils.py to store mysql_connection and slack_client. Tests pass
